rename to graphql-kotlin-spring-server
