--- graphql-kotlin-boot-server/src/main/kotlin/com/expediagroup/graphql/boot/server/annotation/Mutation.kt
@@ -1,22 +1,0 @@
-/*
- * Copyright 2019 Expedia, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.expediagroup.graphql.boot.server.annotation
-
-/**
- * Marker interface to easily autowire all GraphQL mutations.
- */
-interface Mutation
