--- graphql-kotlin-boot-server/src/main/resources/META-INF/spring-configuration-metadata.json
@@ -1,26 +1,0 @@
-{
-  "properties": [
-    {
-      "name": "graphql.endpoint",
-      "type": "java.lang.String",
-      "defaultValue": "graphql",
-      "sourceType": "com.expediagroup.graphql.boot.server.FederationConfigurationProperties"
-    },
-    {
-      "name": "graphql.packages",
-      "type": "java.util.List<java.lang.String>",
-      "sourceType": "com.expediagroup.graphql.boot.server.FederationConfigurationProperties"
-    },
-    {
-      "name": "graphql.federation.enabled",
-      "type": "java.lang.Boolean",
-      "sourceType": "com.expediagroup.graphql.boot.server.FederationConfigurationProperties"
-    },
-    {
-      "name": "graphql.subscriptions.endpoint",
-      "type": "java.lang.String",
-      "defaultValue": "subscriptions",
-      "sourceType": "com.expediagroup.graphql.boot.server.FederationConfigurationProperties"
-    }
-  ]
-}
