--- graphql-kotlin-boot-server/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.expediagroup.graphql.boot.server.GraphQLAutoConfiguration
