+++ graphql-kotlin-schema-generator/src/test/kotlin/com/expediagroup/graphql/generator/extensions/KTypeExtensionsKtTest.kt
@@ -104,8 +104,9 @@ internal class KTypeExtensionsKtTest {
         }
     }
 
-    // with Kotlin 1.3.40+ there is bug on JRE 1.8 that results in incorrect simple name of the anonymous classes
-    // https://youtrack.jetbrains.com/issue/KT-23072
+    // TODO remove JUnit condition once we only build artifacts using Java 11
+    //  BLOCKED: in order to publish to Maven Central we need to generate javadoc jar, blocked until Dokka https://github.com/Kotlin/dokka/issues/294 is resolved
+    //  ISSUE: with Kotlin 1.3.40+ there is bug on JRE 1.8 that results in incorrect simple name of the anonymous classes, see https://youtrack.jetbrains.com/issue/KT-23072
     @Test
     @EnabledOnJre(JRE.JAVA_11)
     fun getSimpleName() {
@@ -116,6 +117,9 @@ internal class KTypeExtensionsKtTest {
         }
     }
 
+    // TODO remove this JUnit once we only build artifacts using Java 11
+    //  BLOCKED: in order to publish to Maven Central we need to generate javadoc jar, blocked until Dokka https://github.com/Kotlin/dokka/issues/294 is resolved
+    //  ISSUE: with Kotlin 1.3.40+ there is bug on JRE 1.8 that results in incorrect simple name of the anonymous classes, see https://youtrack.jetbrains.com/issue/KT-23072
     @Test
     @EnabledOnJre(JRE.JAVA_8)
     fun getSimpleNameJava8() {
