+++ graphql-kotlin-spring-server/src/main/kotlin/com/expediagroup/graphql/spring/server/model/GraphQLRequest.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2019 Expedia, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.expediagroup.graphql.spring.server.model
+
+import graphql.ExecutionInput
+
+data class GraphQLRequest(
+    val query: String,
+    val operationName: String? = null,
+    val variables: Map<String, Any>? = null
+)
+
+fun GraphQLRequest.toExecutionInput(graphQLContext: Any? = null): ExecutionInput =
+    ExecutionInput.newExecutionInput()
+        .query(this.query)
+        .operationName(this.operationName)
+        .variables(this.variables ?: emptyMap())
+        .context(graphQLContext)
+        .build()
