+++ graphql-kotlin-spring-server/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.expediagroup.graphql.spring.server.GraphQLAutoConfiguration
